Messenging-price-analyzer/requirements.txt
# === Core ===
python-dotenv>=1.0.0
pydantic>=2.6
pandas>=2.1.0
requests>=2.31.0
tqdm>=4.66.0
chardet>=5.2.0
protobuf>=5.26.1,<6.0.0

# === Email + attachments parsing ===
mail-parser>=3.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openpyxl>=3.1.2
xlrd>=2.0.0
pdfplumber>=0.11.0
python-docx>=1.0.0

# === LLM (only if used) ===
# google-generativeai>=0.5.0

# === Microsoft Graph (only if used) ===
# msal>=1.28.0

# === ADK + Vertex AI Agent Engine ===
google-adk>=1.15.0,<2.0.0
google-cloud-aiplatform[agent-engines]>=1.118.0,<2.0.0
